# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Bank API
  description: API based on clients from a bank
  termsOfService: http://swagger.io/terms/
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:9091
tags:
  - name: Clients
paths:
  /clients:
    get:
      summary: Get all clients of the bank
      operationId: getAllClients
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clients'
          description: OK
      tags:
        - Clients

    post:
      summary: Save new client
      operationId: saveClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientDTO"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDTO"
      tags:
        - Clients
  /clients/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID from a client
        schema:
          type: string
    get:
      summary: Get a client by ID
      operationId: getClientById
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'
      tags:
        - Clients
    put:
      summary: Update an existing client
      operationId: updateClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientDTO"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDTO"
      tags:
        - Clients
    delete:
      summary: Delete an existing client
      operationId: deleteClient
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDTO"
      tags:
        - Clients
components:
  schemas:
    Clients:
      type: array
      items:
        $ref: "#/components/schemas/ClientDTO"
    ClientDTO:
      title: ClientDTO
      type: object
      properties:
        id:
          description: ID of a client
          type: string
        firstName:
          description: First name of a client
          type: string
          maxLength: 250
        lastName:
          description: Last name of a client
          type: string
          maxLength: 250
        documentType:
          description: Type of documento of a client
          type: string
        documentNumber:
          description: Document number of a client
          type: string
          maxLength: 12
        clientType:
          description: Type of a client (Enterprise or Personal)
          type: string
      required:
        - firstName
        - lastName
        - documentType
        - documentNumber
        - clientType